requirements.txt
Django==2.2.19
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8-plugin-utils==1.3.2
isort==5.10.1
mccabe==0.7.0
Pillow==9.3.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
webcolors==1.12
